+++ src/services/forums/categories.ts
@@ -20,7 +20,7 @@ export async function fetchForumCategories(brand: string): Promise<ForumCategory
 }
 
 export interface CreateForumCategoryParams {
-  name: string
+  title: string
   description: string
   weight: number
   brand: string
@@ -44,7 +44,7 @@ export async function createForumCategory(
 export interface UpdateForumCategoryParams {
   id: number
   brand: string
-  name: string
+  title: string
   weight: number
   description?: string
   icon?: string