+++ src/services/progress-row/method-card.js
@@ -6,7 +6,7 @@ import { getActivePath, fetchLearningPathLessons } from '../content-org/learning
 import { getToday } from '../dateUtils.js'
 import { fetchMethodV2IntroVideo } from '../sanity'
 import { getProgressState } from '../contentProgress'
-import {COLLECTION_TYPE} from "../sync/models/ContentProgress.js";
+import { COLLECTION_TYPE } from '../sync/models/ContentProgress'
 
 export async function getMethodCard(brand) {
   const introVideo = await fetchMethodV2IntroVideo(brand)
@@ -16,7 +16,11 @@ export async function getMethodCard(brand) {
   }
 
   const introVideoProgressState = await getProgressState(introVideo?.id)
-  if (introVideoProgressState !== 'completed') {
+
+  const activeLearningPath = await getActivePath(brand)
+
+  if (introVideoProgressState !== 'completed' || !activeLearningPath) {
+    //startLearningPath('drumeo', 422533)
     const timestamp = Math.floor(Date.now() / 1000)
     const instructorText =
       introVideo.instructor?.length > 1
@@ -40,7 +44,6 @@ export async function getMethodCard(brand) {
     }
   } else {
     //TODO: Optimize loading of dailySessions/Path, should not need multiple requests
-    const activeLearningPath = await getActivePath(brand)
     const learningPath = await fetchLearningPathLessons(
       activeLearningPath.active_learning_path_id,
       brand,
@@ -50,9 +53,11 @@ export async function getMethodCard(brand) {
     const allCompleted = learningPath?.todays_lessons.every(
       (lesson) => lesson.progressStatus === 'completed'
     )
+
     const anyCompleted = learningPath?.todays_lessons.some(
       (lesson) => lesson.progressStatus === 'completed'
     )
+
     const noneCompleted = learningPath?.todays_lessons.every(
       (lesson) => lesson.progressStatus !== 'completed'
     )
@@ -60,6 +65,13 @@ export async function getMethodCard(brand) {
     const nextIncompleteLesson = learningPath?.todays_lessons.find(
       (lesson) => lesson.progressStatus !== 'completed'
     )
+
+    // get the first incomplete lesson from upcoming and next learning path lessons
+    const nextLesson = [
+      ...learningPath?.upcoming_lessons,
+      ...learningPath?.next_learning_path_lessons,
+    ]?.find((lesson) => lesson.progressStatus !== 'completed')
+
     let ctaText, action
     if (noneCompleted) {
       ctaText = 'Start Session'
@@ -68,9 +80,9 @@ export async function getMethodCard(brand) {
       ctaText = 'Continue Session'
       action = getMethodActionCTA(nextIncompleteLesson)
     } else if (allCompleted) {
-      ctaText = learningPath.next_lesson ? 'Start Next Lesson' : 'Browse Lessons'
-      action = learningPath.next_lesson
-        ? getMethodActionCTA(learningPath.next_lesson)
+      ctaText = nextLesson ? 'Start Next Lesson' : 'Browse Lessons'
+      action = nextLesson
+        ? getMethodActionCTA(nextLesson)
         : {
             type: 'lessons',
             brand: brand,
@@ -80,6 +92,7 @@ export async function getMethodCard(brand) {
     let maxProgressTimestamp = Math.max(
       ...learningPath?.children.map((lesson) => lesson.progressTimestamp)
     )
+    
     if (!maxProgressTimestamp) {
       maxProgressTimestamp = learningPath.active_learning_path_created_at
     }